requirements.txt
requests>=2.28.0
pillow>=9.0.0
easyocr>=1.6.0
numpy>=1.20.0
